+++ src/types/youtube.ts
@@ -96,6 +96,9 @@ export class YtChannel {
   // This field serves the purpose of nonce to avoid playback attacks
   lastActedAt: Date
 
+  // Timestamp when the channel verification was processed, either to Verified or Suspended
+  processedAt: Date
+
   // Needs a dummy partition key on GSI to be able to query by createdAt fields
   phantomKey: 'phantomData'
 
@@ -415,8 +418,8 @@ export type TopReferrer = {
 }
 
 export const REFERRAL_REWARD_BY_TIER: { [K in ChannelYppStatusVerified]: number } = {
-  'Bronze': 1,
-  'Silver': 12.5,
-  'Gold': 25,
-  'Diamond': 50,
+  'Bronze': 2,
+  'Silver': 25,
+  'Gold': 50,
+  'Diamond': 100,
 }